Timecode set() fps should be float (from v2 36d6c770082028547da0a6faf02410d858da02c0)

+++ src/wx/timecode.cc
@@ -83,7 +83,7 @@ Timecode::Timecode (wxWindow* parent)
 }
 
 void
-Timecode::set (Time t, int fps)
+Timecode::set (Time t, float fps)
 {
 	/* Do this calculation with frames so that we can round
 	   to a frame boundary at the start rather than the end.
