+++ src/wx/timecode.h
@@ -26,7 +26,7 @@ class Timecode : public wxPanel
 public:
 	Timecode (wxWindow *);
 
-	void set (Time, int);
+	void set (Time, float);
 	Time get (int) const;
 	void clear ();
